+++ types/k6/experimental/timers/index.d.ts
