+++ types/k6/global.d.ts
@@ -56,7 +56,7 @@ declare global {
      * Environment variables.
      * https://grafana.com/docs/k6/latest/using-k6/environment-variables/
      */
-    const __ENV: { [name: string]: string };
+    var __ENV: { [name: string]: string };
 
     // === VU logic only ===
     // ---------------------
@@ -65,11 +65,11 @@ declare global {
      * Current VU number.
      * https://grafana.com/docs/k6/latest/using-k6/execution-context-variables/
      */
-    const __VU: number;
+    var __VU: number;
 
     /**
      * Current iteration number.
      * https://grafana.com/docs/k6/latest/using-k6/execution-context-variables/
      */
-    const __ITER: number;
+    var __ITER: number;
 }
