+++ types/k6/package.json
@@ -2,6 +2,7 @@
     "private": true,
     "name": "@types/k6",
     "version": "0.53.9999",
+    "type": "module",
     "projects": [
         "https://grafana.com/docs/k6/latest/"
     ],
