+++ types/k6/timers/index.d.ts
